refactor(channels): 将 GeminiModelSelection 类型定义移至 types 目录
