+++ src/renderer/pages/settings/channels/components/ChannelPanel.tsx
@@ -7,9 +7,9 @@ import { ConfigStorage } from '@/common/storage';
 import { channel, webui, type IWebUIStatus } from '@/common/ipcBridge';
 import type { IProvider, TProviderWithModel } from '@/common/storage';
 import type { IChannelPluginStatus } from '@/channels/types';
-import { useGeminiModelSelection, type GeminiModelSelection } from '../hooks/useGeminiModelSelection';
+import { useGeminiModelSelection } from '../hooks/useGeminiModelSelection';
 import { useModelProviderList } from '../hooks/useModelProviderList';
-import type { ChannelConfig } from '../types';
+import type { ChannelConfig, GeminiModelSelection } from '../types';
 import { BUILTIN_CHANNEL_TYPES } from '../utils';
 import WeComConfigForm from './WeComConfigForm';
 import WeChatConfigForm from './WeChatConfigForm';
