+++ src/renderer/pages/settings/channels/components/DingTalkConfigForm.tsx
@@ -10,7 +10,7 @@ import { openExternalUrl } from '@/renderer/utils/platform';
 import GeminiModelSelector from '@/renderer/pages/settings/channels/components/GeminiModelSelector';
 import { CHANNEL_DEFAULT_AGENT_BACKEND, type AcpBackendAll } from '@/types/acpTypes';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
-import type { GeminiModelSelection } from '../hooks/useGeminiModelSelection';
+import type { GeminiModelSelection } from '../types';
 import { DINGTALK_DEV_DOCS_URL } from '../utils';
 
 /**
